+++ frontend/src/features/aerodynamics/three/tutorials/hooks/useInduced.ts
